Add basic stimulus controller and ux-turbo to turn the app into a spa
Adds a very basic example of a stimulus controller to the app that makes the sidebar "closable". Apart from that ux-turbo allows the app to basically make ajax calls to the routes that are supposed to be accessed by a full page reload and fetches the html from them, loading it onto the page and therefore not doing a full page reload. This allows for the "speed" and "feel" of a single page application.